requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.9.1
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
charset-normalizer==2.1.1
colorama==0.4.5
colorlog==6.6.0
dnspython==2.2.1
frozenlist==1.3.1
hikari==2.0.0.dev110
hikari-lightbulb==2.2.4
idna==3.3
multidict==6.0.2
pymongo==4.2.0
python-decouple==3.6
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tzdata==2022.2
tzlocal==4.2
yarl==1.8.1
